+++ src/app/_services/compare-auto.service.ts
@@ -1,6 +1,5 @@
 import {Injectable} from '@angular/core';
-import {HttpClient, HttpParams} from '@angular/common/http';
-import { AutoJoin } from '../models/autojoin.model';
+import {HttpClient} from '@angular/common/http';
 import {Observable} from 'rxjs';
 
 @Injectable()
@@ -11,22 +10,18 @@ export class CompareAutoService {
   private compareAutoUrl = 'http://localhost:8080/api/auth';
 
   public getAllAutoToComparePage(params): Observable<any> {
-    console.log("getAllAutoToComparePage -> params: ", params);
-    return this.http.get(this.compareAutoUrl + '/search/compare-auto',{params});
+    return this.http.get(this.compareAutoUrl + '/search/list-compare-auto',{params});
   }
 
   public addAutoToCompare(compareAuto: any):  Observable<any>{
-    console.log("compareAuto: ", compareAuto);
-     return this.http.post(this.compareAutoUrl + '/add-compare-auto', compareAuto);
+     return this.http.post(this.compareAutoUrl + '/add-auto-to-compare', compareAuto);
   }
 
   public deleteCompareAuto(idAuto: any): any{
-    console.log("id auto: ", idAuto);
-    return this.http.delete(this.compareAutoUrl + '/search/compare-auto/'+ idAuto);
+    return this.http.delete(this.compareAutoUrl + '/search/list-compare-auto/'+ idAuto);
   }
 
   public clearListCompareAuto(): any{
-    console.log("clear");
-    return this.http.delete(this.compareAutoUrl + '/search/compare-auto/clear');
+    return this.http.delete(this.compareAutoUrl + '/search/list-compare-auto/clear');
   }
 }
