+++ src/app/_services/token-storage.service.ts
@@ -1,6 +1,4 @@
 import { Injectable } from '@angular/core';
-import {Observable} from "rxjs";
-
 const TOKEN_KEY = 'auth-token';
 const USER_KEY = 'auth-user';
