+++ src/app/_services/user.service.ts
@@ -7,6 +7,7 @@ const API_URL = 'http://localhost:8080/api/auth/';
 @Injectable({
   providedIn: 'root'
 })
+
 export class UserService {
 
   constructor(private http: HttpClient) { }
