+++ src/app/addAuto/add-auto.component.html
@@ -1,5 +1,5 @@
 <div class="content">
-  <h1 class="text-center">Add auto</h1>
+  <h1 class="text-center">Add auto ads</h1>
   <div class="auto-add">
     <div class="container row">
       <div class="col-md-6">
