+++ src/app/app.component.html
@@ -14,6 +14,9 @@
       <li class="nav-item">
         <a style="margin-left:30px" href="/multiple-search-auto" class="nav-link" routerLink="multiple-search-auto">Search auto</a>
       </li>
+      <li class="nav-item">
+        <a style="margin-left:40px" href="/compare-auto" class="nav-link" routerLink="compare-auto">Compare</a>
+      </li>
     </ul>
     <ul class="navbar-nav ml-auto" *ngIf="!isLoggedIn">
       <li class="nav-item">
