+++ src/app/editAuto/edit-auto.component.html
@@ -1,5 +1,5 @@
 <div class="content">
-  <h1>Update auto</h1>
+  <h1>Update auto ads</h1>
   <div class="update-auto">
     <div class="container row">
       <div class="col-md-6">
