+++ src/app/searchAuto/search-auto.component.css
